requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.37.13
botocore==1.37.13
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.0
colorama==0.4.6
cryptography==44.0.2
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.5.0
django-axes==7.0.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-ratelimit==4.1.0
django-storages==1.14.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jmespath==1.0.1
kombu==5.5.0
Markdown==3.8
moviepy==2.1.2
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
pipenv==2024.4.1
pipenv-freeze==0.0.1
platformdirs==4.3.6
proglog==0.1.10
promise==2.3
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.30.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.4
sentry-sdk==2.22.0
setuptools==76.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
text-unidecode==1.3
tomlkit==0.10.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.3
wcwidth==0.2.13
whitenoise==6.9.0
